# api-spec/src/main/resources/static/api.yml
...
swagger: '2.0'
info:
  title: Name of the App API
  description: Use our services using this API
  version: "1.0.0"
# the domain of the service
host: api.app.com
...
schemes:
  - https
# will be prefixed to all paths
basePath: /v1
produces:
  - application/json
paths:
...
  /geofences:
    post:
      summary: Create geofence
      description: |
        This method creates a new geofence.
      parameters:
        - name: geofence_id
          in: query
          description: Unique identifier representing a specific geofence.
          required: true
          type: string
        - name: type_feature
          in: query
          description: Type of the feature.
          required: true
          type: string
        - name: name
          in: query
          description: Name of the geofence.
          required: true
          type: string
        - name: type_geometry
          in: query
          description: Type of the object.
          required: true
          type: string
        - name: coordinates
          in: query
          description: Pair of coordinates describing positions.
          required: true
          type: string
      tags:
        - Geofences
      responses:
        201:
          description: Geofence created
          schema:
            $ref: '#/definitions/Geofence'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    get:
      summary: Get an array of geofences
      description: |
        This method returns an array of those geofences that are inside of an area generated by an specific point and a radius.
      parameters:
        - name: limit
          in: query
          description: Maximum of geofences to be returned
          required: true
          type: integer
        - name: latitude
          in: query
          description: Latitude of a specific point.
          required: true
          type: string
        - name: longitude
          in: query
          description: Longitude of a specific point.
          required: true
          type: string
        - name: radius
          in: query
          description: Distance to determine the area of a circle, taking the specific point as the center of this circle.
          required: true
          type: integer
      tags:
        - Geofences
      responses:
        200:
          description: Geofences inside of the area
          schema:
            type: array
            items:
              $ref: '#/definitions/Geofence'
        404:
          description: Geofences not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /geofences/{geofence_id}:
    put:
      summary: Modify geofence
      description: |
        This method modifies the data of a previously created geofence.
      parameters:
        - name: geofence_id
          in: path
          description: Unique identifier representing a specific geofence.
          required: true
          type: string
        - name: type_feature
          in: query
          description: Type of the feature.
          required: false
          type: string
        - name: name
          in: query
          description: Name of the geofence.
          required: false
          type: string
        - name: type_geometry
          in: query
          description: Type of the object.
          required: false
          type: string
        - name: coordinates
          in: query
          description: Pair of coordinates describing positions.
          required: false
          type: string
      tags:
        - Geofences
      responses:
        202:
          description: Geofence's data modified
          schema:
            $ref: '#/definitions/Geofence'
        304:
          description: Geofence's data not modified
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Geofence not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Delete geofence
      description: |
        This method deletes the data of a previously created geofence.
      parameters:
        - name: geofence_id
          in: path
          description: Unique identifier representing a specific geofence.
          required: true
          type: string
      tags:
        - Geofences
      responses:
        200:
          description: Geofence deleted
          schema:
            $ref: '#/definitions/Geofence'
        404:
          description: Geofence not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    get:
      summary: Get geofence by geofence_id
      description: |
        This method returns a geofence by geofence_id.
      parameters:
        - name: geofence_id
          in: path
          description: Unique identifier representing a specific geofence.
          required: true
          type: string
      tags:
        - Geofences
      responses:
        200:
          description: Geofence with geofence_id requested
          schema:
            $ref: '#/definitions/Geofence'
        404:
          description: Geofence not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Geofence:
    type: object
    properties:
      type:
        type: string
        description: Type of the feature
      geofence_id:
        type: string
        description: Unique identifier representing a specific geofence.
      properties:
        type: object
        properties:
          name:
            type: string
            description: Name of the geofence.
      geometry:
        type: object
        properties:
          type:
            type: string
            description: Type of the object
          coordinates:
            type: array
            items:
              $ref: '#/definitions/Position'
            description: Pair of coordinates describing positions
...
  Position:
    type: object
    properties:
      latitude:
        type: string
        description: Latitude of the point.
      longitude:
        type: string
        description: Longitude of the point.
...
  Error:
    type: object
    properties:
      title:
        type: string
      body:
        type: string